requirements.txt
# Core ML Libraries
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0

# API & Serving
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Experiment Tracking (optional - for MLOps phase)
# mlflow>=2.9.0

# Utilities
pyyaml>=6.0.0
tqdm>=4.66.0
python-dotenv>=1.0.0

# Jupyter
jupyter>=1.0.0
ipywidgets>=8.1.0

# Testing
pytest>=7.4.0

# Code Quality
black>=23.0.0
isort>=5.12.0
